pyproject.toml
[tool.poetry]
name = "py-hometask"
version = "0.1.0"
description = ""
authors = ["Max Kobozev <quesesr@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.6"
pyright = "^1.1.375"
pytest = "^8.3.2"

[tool.ruff]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
